flask_nocobase_importer/requirements.txt
Flask
psycopg2-binary
pandas
openpyxl
pyarrow
numpy
